moved some css around
