+++ src/assets/images/gdsc-logo-and-text.png
