+++ src/index.scss
@@ -1,4 +1,4 @@
-@import "pages/login-page/login";
+@import "styles/pages/login";
 
 /*For newly created css files you need to import them here*/
