+++ src/styles/general/App.scss
@@ -1,4 +1,4 @@
-@import "styles/_font.scss";
+@import "font";
 
 .App {
     text-align: center;
