+++ src/styles/pages/login.scss
@@ -25,11 +25,6 @@ Form {
   background-color: #fff;
 }
 
-.header {
-  padding-left: 3em;
-  padding-right: 3em;
-}
-
 img {
   max-width: 400px;
   width: 100%;
